Fixing whitespace crash
Fixing the whitespace crash which was caused by an oversight in the command evaluation which then led to an ArrayIndexOutOfBound exception.

The exception was actually handled, yet for some reason the app was instructed to shut down in this case.

In conclusion, this was probably a leftover from early development.